Created initial code. It works for the most part and currently outputs to the console for debugging purposes.
The only thing left for class creation is the database interaction. There is an uncomplete method called database_lookup_type that will be given a column name. We need to access the mysql information schema table and do a lookup on the column. After that, we need to determine what the corresponding C# type is.